renamed handeye_simulator
